delete react-ibsheet
